Add the grouping option to the component wizard
The user can now choose if the uses/provides entries are to be grouped after
the component creation by the component wizard. If the user chooses so, all
uses entries and provides entries are grouped and written together and the
keywords 'uses' and 'provides' appear at most once in the generated code.