Add project preferences for comments
A page with project specific preferences for comments is added. For each
project and each comment type the user can choose one of three options: to
inherit the comment from the global preferences, not to generate the comment of
this type or to enter a project specific comment.
Moreover, objects used for generating initial contents of files created by
wizards are changed. Now, character streams are used (they replaced byte
streams). It fixes computing the initial position of the cursor after a file
creation - it was sometimes invalid.